Allow watchdog grace period to reboot before "kicking" watchdog reboot with software reboot.

+++ pkg/reboot/rebooter.go
@@ -2,6 +2,8 @@ package reboot
 
 import (
 	"os/exec"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/medik8s/self-node-remediation/pkg/watchdog"
@@ -9,7 +11,7 @@ import (
 
 type Rebooter interface {
 	// Reboot triggers a node reboot
-	Reboot() error
+	Reboot() (ctrl.Result, error)
 }
 
 var _ Rebooter = &WatchdogRebooter{}
@@ -27,27 +29,47 @@ func NewWatchdogRebooter(wd watchdog.Watchdog, log logr.Logger) Rebooter {
 	}
 }
 
-func (r *WatchdogRebooter) Reboot() error {
+func (r *WatchdogRebooter) Reboot() (ctrl.Result, error) {
+	if r.isWatchdogRebootStuck() {
+		return r.softwareReboot()
+	}
+	//Watch dog is rebooting, wait to make sure watchdog is rebooting properly otherwise intervene with software reboot
+	if r.wd.IsRebooting() {
+		r.log.Info("watchdog is waiting for reboot to commence, waiting for 10 seconds")
+		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
+	}
+
 	if r.wd == nil || !r.wd.IsStarted() {
 		r.log.Info("no watchdog is present on this host, trying software reboot")
 		//we couldn't init a watchdog so far but requested to be rebooted. we issue a software reboot
-		if err := r.softwareReboot(); err != nil {
-			r.log.Error(err, "failed to run reboot command")
-			// TODO retry because of this?
-			//return err
-			return nil
-		}
-		return nil
+		return r.softwareReboot()
 	}
 	// we stop feeding the watchdog for a reboot
 	r.wd.Stop()
 	r.log.Info("watchdog feeding has stopped, waiting for reboot to commence")
-	return nil
+	return ctrl.Result{RequeueAfter: time.Second * 10}, nil
 }
 
 // softwareReboot performs software reboot by running systemctl reboot
-func (r *WatchdogRebooter) softwareReboot() error {
+func (r *WatchdogRebooter) softwareReboot() (ctrl.Result, error) {
 	// hostPID: true and privileged:true required to run this
 	rebootCmd := exec.Command("/usr/bin/nsenter", "-m/proc/1/ns/mnt", "/bin/systemctl", "reboot", "--force", "--force")
-	return rebootCmd.Run()
+
+	if err := rebootCmd.Run(); err != nil {
+		r.log.Error(err, "failed to run reboot command")
+		// TODO retry because of this?
+	}
+	return ctrl.Result{}, nil
+}
+
+func (r *WatchdogRebooter) isWatchdogRebootStuck() bool {
+	timeToAssumeRebootHasStarted := time.Second * 30
+	lastFoodTime := r.wd.LastFoodTime()
+	var isStuck bool
+	now := time.Now()
+	if isStuck = now.After(lastFoodTime.Add(timeToAssumeRebootHasStarted)); isStuck {
+		r.log.Info("watchdog reboot is stuck, about to try software reboot")
+	}
+
+	return isStuck
 }
